Crowd Counting based on density
Once the above model is ready we can give an input image and based on the density estimates given by all three models we will be able to approximate, if the crowd is in hundreds or thousands. Along with this, we will also be able to recognize the highly populated places in that particular image. This result can be achieved even on a freshly fed image of random resolution with decent quality.